[TEST][#59] Move test from Client to Compiler [#950]
