+++ Compiler/TranslatorTests/TestProjects/16/Bridge/reference/test.bridgeIssues.n059.js
@@ -1,0 +1,20 @@
+﻿/* global Bridge */
+
+(function (globals) {
+    "use strict";
+
+    Bridge.define('Test.BridgeIssues.N059.Class59', {
+        statics: {
+            method1: function () {
+            }
+        },
+        method1: function (d) {
+        }
+    });
+    
+    Bridge.define('Test.BridgeIssues.N059.Class59.Aux1');
+    
+    
+    
+    Bridge.init();
+})(this);
