+++ Compiler/TranslatorTests/TestProjects/16/BridgeIssues/N059.cs
@@ -1,0 +1,20 @@
+﻿using Bridge;
+
+namespace Test.BridgeIssues.N059
+{
+    // Does not compile issue [#59]
+    internal class Class59
+    {
+        public static void Method1()
+        {
+        }
+
+        public void Method1(Aux1 d)
+        {
+        }
+
+        public class Aux1
+        {
+        }
+    }
+}
